# docs/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "AssessorLocation": {
        "additionalProperties": false,
        "properties": {
          "region": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Region",
            "type": "string"
          }
        },
        "required": [
          "region"
        ],
        "title": "AssessorLocation",
        "type": "object"
      },
      "AssessorRoutingResult": {
        "additionalProperties": false,
        "properties": {
          "assessor_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessor Reference"
          },
          "expected_by": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected By"
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "next_step": {
            "title": "Next Step",
            "type": "string"
          },
          "queue_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Queue Reference"
          },
          "routing_status": {
            "$ref": "#/components/schemas/AssessorRoutingStatus"
          }
        },
        "required": [
          "routing_status",
          "next_step"
        ],
        "title": "AssessorRoutingResult",
        "type": "object"
      },
...
      "ClaimHistorySearchRequest": {
        "additionalProperties": false,
        "description": "Purpose-limited history lookup.\n\n`purpose` is an allow-list rather than free text so a caller cannot widen\nthe reason for reading a claimant's history.",
        "properties": {
          "claim_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Claim Id",
            "type": "string"
          },
          "history_reference": {
            "maxLength": 200,
            "minLength": 1,
            "title": "History Reference",
            "type": "string"
          },
          "limit": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "purpose": {
            "const": "relevant_history_review",
            "default": "relevant_history_review",
            "title": "Purpose",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "history_reference"
        ],
        "title": "ClaimHistorySearchRequest",
        "type": "object"
      },
      "ClaimHistorySearchResponse": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ClaimHistoryFacts"
              },
              {
                "type": "null"
              }
            ]
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Limitations",
            "type": "array"
          },
          "result_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Result Id",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/RetrievalStatus"
          },
          "uncertainty": {
            "items": {
              "$ref": "#/components/schemas/RetrievalUncertainty"
            },
            "maxItems": 100,
            "title": "Uncertainty",
            "type": "array"
          }
        },
        "required": [
          "result_id",
          "status",
          "retrieved_at"
        ],
        "title": "ClaimHistorySearchResponse",
        "type": "object"
      },
...
      "CompleteEvidenceProcessingRequest": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/EvidenceFactProposal"
            },
            "maxItems": 50,
            "minItems": 1,
            "title": "Facts",
            "type": "array"
          }
        },
        "required": [
          "facts"
        ],
        "title": "CompleteEvidenceProcessingRequest",
        "type": "object"
      },
...
      "CreateExternalClaimRequest": {
        "additionalProperties": false,
        "properties": {
          "authorised_decision_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Authorised Decision Id",
            "type": "string"
          },
          "claim_revision": {
            "minimum": 1.0,
            "title": "Claim Revision",
            "type": "integer"
          },
          "confirmed_form": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StructuredFormField-Input"
            },
            "title": "Confirmed Form",
            "type": "object"
          },
          "evidence_refs": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Evidence Refs",
            "type": "array"
          },
          "pending_evidence": {
            "items": {
              "$ref": "#/components/schemas/PendingEvidenceReference"
            },
            "maxItems": 100,
            "title": "Pending Evidence",
            "type": "array"
          },
          "route": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Route",
            "type": "string"
          },
          "working_claim_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Working Claim Id",
            "type": "string"
          }
        },
        "required": [
          "working_claim_id",
          "claim_revision",
          "authorised_decision_id",
          "route"
        ],
        "title": "CreateExternalClaimRequest",
        "type": "object"
      },
...
      "EvidenceProcessingResponse": {
        "additionalProperties": false,
        "properties": {
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "file_status": {
            "$ref": "#/components/schemas/EvidenceFileStatus"
          },
          "proposed_fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/StructuredFormField-Output"
            },
            "title": "Proposed Fields",
            "type": "object"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          }
        },
        "required": [
          "evidence_id",
          "revision",
          "file_status",
          "proposed_fields"
        ],
        "title": "EvidenceProcessingResponse",
        "type": "object"
      },
...
      "ExternalClaimResult": {
        "additionalProperties": false,
        "properties": {
          "claim_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Claim Number"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "creation_status": {
            "$ref": "#/components/schemas/ClaimCreationStatus"
          },
          "expected_by": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected By"
          },
          "external_claim_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "External Claim Id"
          },
          "next_step": {
            "title": "Next Step",
            "type": "string"
          },
          "route": {
            "title": "Route",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/IntegrationSource"
          }
        },
        "required": [
          "creation_status",
          "route",
          "next_step",
          "source",
          "created_at"
        ],
        "title": "ExternalClaimResult",
        "type": "object"
      },
...
      "ExternalTaskListResponse": {
        "additionalProperties": false,
        "description": "Internal projection of external tasks and their material links.",
        "properties": {
          "claim_id": {
            "title": "Claim Id",
            "type": "string"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExternalTaskEvidenceView"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "$ref": "#/components/schemas/PageInfo"
          }
        },
        "required": [
          "claim_id",
          "items",
          "page"
        ],
        "title": "ExternalTaskListResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "KnowledgeSearchRequest": {
        "additionalProperties": false,
        "properties": {
          "authority": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Authority",
            "type": "string"
          },
          "document_id": {
            "anyOf": [
              {
                "maxLength": 200,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Id"
          },
          "effective_at": {
            "format": "date-time",
            "title": "Effective At",
            "type": "string"
          },
          "insurer": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Insurer",
            "type": "string"
          },
          "jurisdiction": {
            "maxLength": 20,
            "minLength": 2,
            "title": "Jurisdiction",
            "type": "string"
          },
          "limit": {
            "default": 5,
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "product": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Product",
            "type": "string"
          },
          "question": {
            "maxLength": 500,
            "minLength": 1,
            "title": "Question",
            "type": "string"
          },
          "version": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Version",
            "type": "string"
          },
          "visibility": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Visibility",
            "type": "string"
          }
        },
        "required": [
          "question",
          "jurisdiction",
          "visibility",
          "authority",
          "version",
          "insurer",
          "product",
          "effective_at"
        ],
        "title": "KnowledgeSearchRequest",
        "type": "object"
      },
      "KnowledgeSearchResponse": {
        "additionalProperties": false,
        "properties": {
          "limitations": {
            "items": {
              "type": "string"
            },
            "title": "Limitations",
            "type": "array"
          },
          "results": {
            "items": {
              "$ref": "#/components/schemas/KnowledgeCitation"
            },
            "title": "Results",
            "type": "array"
          },
          "status": {
            "enum": [
              "evidence_found",
              "no_evidence",
              "unavailable"
            ],
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "status",
          "results",
          "limitations"
        ],
        "title": "KnowledgeSearchResponse",
        "type": "object"
      },
...
      "PolicySearchRequest": {
        "additionalProperties": false,
        "properties": {
          "claim_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Claim Id",
            "type": "string"
          },
          "effective_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective At"
          },
          "policy_reference": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Policy Reference",
            "type": "string"
          },
          "question": {
            "anyOf": [
              {
                "maxLength": 500,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Question"
          }
        },
        "required": [
          "claim_id",
          "policy_reference"
        ],
        "title": "PolicySearchRequest",
        "type": "object"
      },
      "PolicySearchResponse": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PolicyFacts"
              },
              {
                "type": "null"
              }
            ]
          },
          "limitations": {
            "items": {
              "type": "string"
            },
            "maxItems": 100,
            "title": "Limitations",
            "type": "array"
          },
          "result_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Result Id",
            "type": "string"
          },
          "retrieved_at": {
            "format": "date-time",
            "title": "Retrieved At",
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RetrievalSource"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/RetrievalStatus"
          },
          "uncertainty": {
            "items": {
              "$ref": "#/components/schemas/RetrievalUncertainty"
            },
            "maxItems": 100,
            "title": "Uncertainty",
            "type": "array"
          }
        },
        "required": [
          "result_id",
          "status",
          "retrieved_at"
        ],
        "title": "PolicySearchResponse",
        "type": "object"
      },
...
      "RouteAssessorRequest": {
        "additionalProperties": false,
        "properties": {
          "authorisation_ref": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Authorisation Ref",
            "type": "string"
          },
          "claim_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Claim Id",
            "type": "string"
          },
          "claimant_consent_ref": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Claimant Consent Ref",
            "type": "string"
          },
          "external_claim_id": {
            "maxLength": 100,
            "minLength": 1,
            "title": "External Claim Id",
            "type": "string"
          },
          "location": {
            "$ref": "#/components/schemas/AssessorLocation"
          },
          "requested_action": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Requested Action",
            "type": "string"
          }
        },
        "required": [
          "claim_id",
          "external_claim_id",
          "authorisation_ref",
          "claimant_consent_ref",
          "requested_action",
          "location"
        ],
        "title": "RouteAssessorRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Northwind FNOL Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/internal/v1/assessors/route": {
      "post": {
        "operationId": "route_assessor_integration_internal_v1_assessors_route_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteAssessorRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssessorRoutingResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Route Assessor Integration",
        "tags": [
          "internal-integrations"
        ]
      }
    },
    "/internal/v1/claim-history/search": {
      "post": {
        "operationId": "search_claim_history_integration_internal_v1_claim_history_search_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClaimHistorySearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClaimHistorySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Claim History Integration",
        "tags": [
          "internal-integrations"
        ]
      }
    },
    "/internal/v1/claims/create": {
      "post": {
        "operationId": "create_claim_integration_internal_v1_claims_create_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateExternalClaimRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalClaimResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Claim Integration",
        "tags": [
          "internal-integrations"
        ]
      }
    },
    "/internal/v1/claims/{claim_id}/evidence/{evidence_id}/processing": {
      "post": {
        "operationId": "complete_evidence_processing_integration_internal_v1_claims__claim_id__evidence__evidence_id__processing_post",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "evidence_id",
            "required": true,
            "schema": {
              "title": "Evidence Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "If-Match",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "If-Match"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteEvidenceProcessingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceProcessingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Evidence Processing Integration",
        "tags": [
          "internal-integrations"
        ]
      }
    },
    "/internal/v1/claims/{claim_id}/external-tasks": {
      "get": {
        "operationId": "read_external_tasks_integration_internal_v1_claims__claim_id__external_tasks_get",
        "parameters": [
          {
            "in": "path",
            "name": "claim_id",
            "required": true,
            "schema": {
              "title": "Claim Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 25,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalTaskListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Read External Tasks Integration",
        "tags": [
          "internal-integrations"
        ]
      }
    },
    "/internal/v1/knowledge/search": {
      "post": {
        "operationId": "search_knowledge_integration_internal_v1_knowledge_search_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/KnowledgeSearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeSearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Knowledge Integration",
        "tags": [
          "internal-integrations"
        ]
      }
    },
    "/internal/v1/policy/search": {
      "post": {
        "operationId": "search_policy_integration_internal_v1_policy_search_post",
        "parameters": [
          {
            "in": "header",
            "name": "authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PolicySearchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicySearchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Policy Integration",
        "tags": [
          "internal-integrations"
        ]
      }
    }
  }
}